python-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
langdetect==1.0.9
jieba==0.42.1
nltk==3.8.1
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
redis==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
